fix wf
