+++ graph_core/tests/path.yaml
@@ -1,18 +1,16 @@
 graph_core_tests:
   path:
     - [-1.5, -1.5, -1.5]
-    - [-1.289048327034827, -0.8897895079058549, -0.5725643821398563]
-    - [-1.202258162061804, -0.3845433671421106, 0.05580644475100871]
-    - [-1.124060821544408, -0.2976412851352439, 0.2980147175481612]
-    - [-1.111268186295064, -0.09690538635333312, 0.6861515521007364]
-    - [-1.003801554210392, 0.07328138108052384, 1.030200171996932]
-    - [-0.63156594202461, 0.2985122789621875, 1.092909070939249]
-    - [-0.1655723749033977, 0.5608904445827428, 1.211830181401144]
-    - [0.1298909716447308, 0.6759194695278627, 1.271348150340537]
-    - [0.2381955309483202, 0.7555878084411788, 1.313497947442112]
-    - [0.5706435502835752, 0.8809782952912981, 1.292315213378666]
-    - [0.6681131330170265, 0.9888370397914372, 1.304449596816883]
-    - [0.7749782680882972, 1.029822015450253, 1.324334162671275]
-    - [1.032189844889655, 1.156205783438033, 1.417469777594074]
-    - [1.268195718884559, 1.329730769540058, 1.450277447677347]
+    - [-1.368304943282731, -0.869310126335348, -0.7325118657666546]
+    - [-1.113252883580279, -0.09815786150562764, 0.5658478923234942]
+    - [-0.986172691213979, 0.05930640853909153, 1.050255684438281]
+    - [-0.7429864148673528, 0.2486760979916678, 1.160513419597364]
+    - [-0.2367755061652399, 0.5426347130176773, 1.333123798870073]
+    - [0.221342196558296, 0.8029216822716043, 1.401807888830923]
+    - [0.3617322458241751, 0.8665115658037883, 1.374848947802023]
+    - [0.5586325554450194, 1.018791281859759, 1.367347763789514]
+    - [0.6556395875083467, 1.103812723236071, 1.374659214110421]
+    - [0.8294250552213867, 1.236146434087375, 1.394957294173053]
+    - [1.024025496339438, 1.352149415226723, 1.435204434178399]
+    - [1.284559647918008, 1.429619306665668, 1.478383608152337]
     - [1.5, 1.5, 1.5]
